Change the minSdkVersion back to 14, addition of some javadoc, change the (weird) packaging, split some long methods into several shorter methods
