+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/coin/coins/GoldCoin.java
@@ -1,8 +1,8 @@
-package com.context.kroket.escapeapp.minigames.minigames.coins;
+package com.context.kroket.escapeapp.minigames.coin.coins;
 
 import android.widget.ImageView;
 
-import com.context.kroket.escapeapp.minigames.minigames.coin.Coin;
+import com.context.kroket.escapeapp.minigames.coin.Coin;
 
 /**
  * Created by Team Kroket on 5-6-2016.
